requirements.txt
google-generativeai
numpy
pynput
python-dotenv
requests
sounddevice
soundfile
torch
transformers